Update schwer-scripts
Match itsschwer/schwer-scripts@c1de9a1

- Ignore demo prefab changes (modified in project)
