+++ Assets/schwer-scripts-master/schwer-scripts/Editor/ScriptableObjectUtility.cs
@@ -4,11 +4,11 @@ using UnityEngine;
 
 namespace SchwerEditor {
     /// <summary>
-    /// Editor class for working with Scriptable Object assets.
+    /// An editor-only class for working with <c>ScriptableObject</c> assets.
     /// </summary>
     public static class ScriptableObjectUtility {
         /// <summary>
-        /// Creates a new Scriptable Object asset in a manner that mimics Unity's asset creation process.
+        /// Creates a new <c>ScriptableObject</c> asset in a manner that mimics Unity's asset creation process.
         /// </summary>
         public static T CreateAsset<T>() where T : ScriptableObject {
             // From: https://wiki.unity3d.com/index.php/CreateScriptableObjectAsset
@@ -22,7 +22,7 @@ namespace SchwerEditor {
                 path = path.Replace(Path.GetFileName(AssetDatabase.GetAssetPath(Selection.activeObject)), "");
             }
             path = AssetDatabase.GenerateUniqueAssetPath(path + "/" + typeof(T).Name + ".asset");
-            
+
             AssetDatabase.CreateAsset(asset, path);
             AssetsUtility.SaveRefreshAndFocus();
             return asset;
