+++ Assets/schwer-scripts-master/schwer-scripts/ItemSystem/Editor/InventoryDrawer.cs
@@ -18,7 +18,7 @@ namespace SchwerEditor.ItemSystem {
             }
 
             var foldoutRect = new Rect(position.x, position.y, position.width, EditorGUIUtility.singleLineHeight);
-            property.isExpanded = EditorGUI.Foldout(foldoutRect, property.isExpanded, "Contents (" + keys.arraySize + ")", true);
+            property.isExpanded = EditorGUI.Foldout(foldoutRect, property.isExpanded, $"{GetDisplayName(property)} ({keys.arraySize})", true);
 
             if (property.isExpanded) {
                 EditorGUI.BeginDisabledGroup(true);
@@ -40,6 +40,14 @@ namespace SchwerEditor.ItemSystem {
 
                 EditorGUI.EndDisabledGroup();
             }
+        }
+
+        private string GetDisplayName(SerializedProperty property) {
+            // Personal practice when using ScriptableObjects that act as an
+            // SO container for a plain C# type is to name the member `value`.
+            // However, I want `InventorySO.value` to display as 'Contents'.
+            var objectType = property.serializedObject.targetObject.GetType();
+            return (objectType == typeof(InventorySO)) ? "Contents" : property.displayName;
         }
     }
 }
