+++ Assets/schwer-scripts-master/schwer-scripts/ItemSystem/Editor/ItemDatabaseInspector.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 87db772437a93b34d9778f86b040c313
+guid: e6bd7579005c56449b50c1312022b4b5
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
