+++ Assets/schwer-scripts-master/schwer-scripts/ScriptableDatabase.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 3e0f4be3b4eb5a047840246e5d0a47f2
+guid: 4c17e809a4799694f837a9387ab5cfad
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
