+++ Assets/schwer-scripts-master/schwer-scripts/ScriptableDatabase/Editor.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: dd45106e979b4714e83d5424a6b71ca1
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
