+++ Assets/schwer-scripts-master/schwer-scripts/ScriptableDatabase/Editor/ScriptableDatabaseInspector.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: ba2f7e4b24526451587749862881fdc8
+guid: 795dbd3566a5ab04fb3ad4a8ebbd8576
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
