+++ Assets/schwer-scripts-master/schwer-scripts/ScriptableDatabase/Editor/ScriptableDatabaseUtility.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: c9f51e0a1a1e06548b6da9fa7a0a27d0
+guid: f5505e2350674ef41872f9d62c22cd5e
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
