+++ Assets/schwer-scripts-master/schwer-scripts/ScriptableDatabase/README.md.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: faf720060cde118469c3529ceca1d45e
-folderAsset: yes
-DefaultImporter:
+guid: 5c3c37a177078724e851181a4b8c28d3
+TextScriptImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
