+++ Assets/schwer-scripts-master/schwer-scripts/ScriptableDatabase/ScriptableDatabase.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: abeb412bcae714e4780959380f287828
+guid: acd4a96a23def9949905cd7cf8463513
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
